requirements.txt
starlette
uvicorn
fastapi
SQLAlchemy
jinja2
python-multipart
pydantic
psycopg2-binary
